Qute - get rid of VariantTemplate and unify func. of injected templates
- resolves #8754
