+++ extensions/qute/deployment/src/test/java/io/quarkus/qute/deployment/VariantTemplateTest.java
@@ -11,9 +11,9 @@ import org.jboss.shrinkwrap.api.spec.JavaArchive;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.RegisterExtension;
 
+import io.quarkus.qute.Template;
 import io.quarkus.qute.TemplateInstance;
 import io.quarkus.qute.Variant;
-import io.quarkus.qute.api.VariantTemplate;
 import io.quarkus.test.QuarkusUnitTest;
 
 public class VariantTemplateTest {
@@ -31,9 +31,9 @@ public class VariantTemplateTest {
     @Test
     public void testRendering() {
         TemplateInstance rendering = simpleBean.foo.instance().data("bar");
-        rendering.setAttribute(VariantTemplate.SELECTED_VARIANT, new Variant(null, "text/plain", null));
+        rendering.setAttribute(TemplateInstance.SELECTED_VARIANT, new Variant(null, "text/plain", null));
         assertEquals("bar", rendering.render());
-        rendering.setAttribute(VariantTemplate.SELECTED_VARIANT, new Variant(null, "text/html", null));
+        rendering.setAttribute(TemplateInstance.SELECTED_VARIANT, new Variant(null, "text/html", null));
         assertEquals("<strong>bar</strong>", rendering.render());
     }
 
@@ -41,7 +41,7 @@ public class VariantTemplateTest {
     public static class SimpleBean {
 
         @Inject
-        VariantTemplate foo;
+        Template foo;
 
     }
